backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
aiokafka==0.10.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
prometheus-client==0.21.1
python-json-logger==2.0.7
PyYAML==6.0.2
kubernetes==31.0.0
ansible-runner==2.4.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
pytest==8.4.1
